Revert "refactor!: Better Integer handling"
This reverts commit 7271a4ff86e59d04ab73706581b86a36e82ea627.

+++ frappe/database/mariadb/schema.py
@@ -52,8 +52,8 @@ class MariaDBTable(DBTable):
 			modified datetime(6),
 			modified_by varchar({varchar_len}),
 			owner varchar({varchar_len}),
-			docstatus tinyint not null default '0',
-			idx int not null default '0',
+			docstatus int(1) not null default '0',
+			idx int(8) not null default '0',
 			{additional_definitions})
 			ENGINE={engine}
 			ROW_FORMAT=DYNAMIC
